Add Authentication class with TestLDAPConnection class
